[hotfix][tests] Extract AcknowledgeStreamMockEnvironment
